backend/requirements.txt
# In facials/backend/requirements.txt

# Core Django
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary

# Authentication
djoser
djangorestframework-simplejwt

# Real-Time / ASGI
channels>=4.0,<5.0
channels-redis>=4.0,<5.0
daphne>=4.0,<5.0

# Face Recognition
deepface
Pillow
opencv-python-headless

# Utilities
django-cors-headers
openpyxl